Add span ops and move things around
