+++ relay-server/src/endpoints/common.rs
@@ -2,7 +2,6 @@
 
 use axum::http::{header, StatusCode};
 use axum::response::IntoResponse;
-use bytes::Bytes;
 use relay_general::protocol::{EventId, EventType};
 use relay_quotas::RateLimits;
 use relay_statsd::metric;
@@ -143,25 +142,6 @@ impl IntoResponse for BadStoreRequest {
         }
 
         response
-    }
-}
-
-/// Adds the instrumentation layer on top of the [`bytes::Bytes`].
-#[derive(Debug)]
-pub(crate) struct BytesWrapper(pub(crate) Bytes);
-
-#[axum::async_trait]
-impl<S, B> axum::extract::FromRequest<S, B> for BytesWrapper
-where
-    Bytes: axum::extract::FromRequest<S, B>,
-    B: Send + 'static,
-    S: Send + Sync,
-{
-    type Rejection = <Bytes as axum::extract::FromRequest<S, B>>::Rejection;
-
-    #[tracing::instrument(name = "Bytes::from_request", skip_all)]
-    async fn from_request(req: axum::http::Request<B>, state: &S) -> Result<Self, Self::Rejection> {
-        Ok(Self(Bytes::from_request(req, state).await?))
     }
 }
 
@@ -342,7 +322,7 @@ fn queue_envelope(
 ///
 /// This returns `Some(EventId)` if the envelope contains an event, either explicitly as payload or
 /// implicitly through an item that will create an event during ingestion.
-#[tracing::instrument(skip_all)]
+#[tracing::instrument(name = "function", skip_all)]
 pub async fn handle_envelope(
     state: &ServiceState,
     envelope: Box<Envelope>,
@@ -370,7 +350,7 @@ pub async fn handle_envelope(
     let checked = state
         .project_cache()
         .send(CheckEnvelope::new(managed_envelope))
-        .instrument(tracing::debug_span!("ProjectCache::CheckEnvelope"))
+        .instrument(tracing::debug_span!("queue.submit"))
         .await
         .map_err(|_| BadStoreRequest::ScheduleFailed)?
         .map_err(BadStoreRequest::EventRejected)?;
