+++ relay-server/src/extractors/request_meta.rs
@@ -426,7 +426,7 @@ where
 {
     type Rejection = Infallible;
 
-    #[tracing::instrument(skip_all, name = "PartialMeta::from_request")]
+    #[tracing::instrument(name = "middleware.axum.extractor", skip_all)]
     async fn from_request_parts(parts: &mut Parts, state: &S) -> Result<Self, Self::Rejection> {
         let mut ua = RawUserAgentInfo::default();
         for (key, value) in &parts.headers {
@@ -545,7 +545,7 @@ struct StorePath {
 impl FromRequestParts<ServiceState> for RequestMeta {
     type Rejection = BadEventMeta;
 
-    #[tracing::instrument(skip_all, name = "RequestMeta::from_request")]
+    #[tracing::instrument(name = "middleware.axum.extractor", skip_all)]
     async fn from_request_parts(
         parts: &mut Parts,
         state: &ServiceState,
