+++ relay-server/src/extractors/signed_json.rs
@@ -81,7 +81,7 @@ where
 {
     type Rejection = SignatureError;
 
-    #[tracing::instrument(skip_all)]
+    #[tracing::instrument(name = "middleware.axum.extractor", skip_all)]
     async fn from_request(
         request: Request<B>,
         state: &ServiceState,
